First Commit Late Start Here we Go
Im sorry I keep starting these projects late, but Saturday through Monday seems to be the best block of time for me to get stuff done. I scrapped together a previous assignment to bring me this. Luckily theres still lots of work to do revising this project.  I need to get the ajax calls to work by referencing the right database tables, then I'm going to have to modify the handlebars to change the layout so each burger has it's own button, and make the eaten burger list. Then I'm going to figure out how to deploy this on heroku.